Folder structure creation and code separation
